Create user routes
Added login, signup, and current user routes. '/register' checks
for already existing db entries, and uses BCrypt to hash the password
field of a user.

'/login' uses BCrypt to compare the passwords and uses the jwt module
to sign the payload (consisting of the email and password) and uses
a secretOrKey to sign the payload.

Throughout all routes, errors are being sent as responses if the user
fails any model constraints.